# api/API/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de blog du cours Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "sandbox-server"
        }
    ],
    "paths": {
...
        "/don": {
            "get": {
                "summary": "List all donations",
                "description": "List all donations",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit permettant la pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all donations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Don[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Don"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/don/{id}": {
            "get": {
                "summary": "return a donation",
                "description": "return a donation",
                "operationId": "2359083d99f65aeb11c69d08d6c3bc6a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du don à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return a donation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Don"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/don/create": {
            "post": {
                "summary": "Create a donation",
                "description": "Save donation in DB",
                "operationId": "aa748e7fd4ae5367a067fafba62113da",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Don à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pseudo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "montant": {
                                        "type": "double",
                                        "example": 50.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Don enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/don/{id}/update": {
            "put": {
                "summary": "Update donation",
                "description": "Update donation in Db",
                "operationId": "d311a28bb723860b4767ce20e8bbec49",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de don à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nouvelles données du don",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pseudo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "montant": {
                                        "type": "double",
                                        "example": 40.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Don enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/don/{id}/delete": {
            "delete": {
                "summary": "Delete donation",
                "description": "Delete donation in Db",
                "operationId": "0ef5832b6feb1196eb78fd5cfd051394",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du don à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Don supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    }
}